chore: cont-opt[2] enable dumping witness table periodically
