+++ crates/specs/src/configure_table.rs
@@ -6,7 +6,7 @@ pub const WASM_BYTES_PER_PAGE: u64 = 64 * 1024 as u64;
 
 const WASM_32_MAXIMAL_PAGES_DEFAULT: u32 = 65536;
 
-#[derive(Serialize, Debug, Clone, Copy, Deserialize)]
+#[derive(Serialize, Debug, Clone, Copy, Deserialize, PartialEq)]
 pub struct ConfigureTable {
     pub init_memory_pages: u32,
     pub maximal_memory_pages: u32,
