+++ crates/zkwasm/Cargo.toml
@@ -35,4 +35,4 @@ rusty-fork = "0.3.0"
 [features]
 default = []
 cuda = ["halo2_proofs/cuda", "specs/cuda"]
-continuation = ["specs/continuation"]
+continuation = ["specs/continuation", "wasmi/continuation"]
