+++ crates/zkwasm/src/circuits/utils/table_entry.rs
@@ -5,6 +5,8 @@ use specs::mtable::AccessType;
 use specs::mtable::LocationType;
 use specs::mtable::MTable;
 use specs::mtable::MemoryTableEntry;
+use specs::imtable::memory_event_of_step;
+
 use std::cmp::Ordering;
 use std::collections::BTreeMap;
 use std::env;
@@ -12,7 +14,7 @@ use std::io::Write;
 use std::path::PathBuf;
 
 use crate::circuits::config::zkwasm_k;
-use crate::runtime::memory_event_of_step;
+
 
 #[derive(Clone, Debug, Serialize)]
 pub(in crate::circuits) struct MemoryWritingEntry {
