+++ crates/zkwasm/src/runtime/wasmi_interpreter.rs
@@ -5,18 +5,18 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use specs::host_function::HostFunctionDesc;
-use specs::jtable::StaticFrameEntry;
-use specs::state::InitializationState;
+use specs::state::{InitializationState, UpdateCompilationTable};
 use specs::CompilationTable;
 use specs::ExecutionTable;
 use specs::Tables;
+
 use wasmi::Externals;
 use wasmi::ImportResolver;
 use wasmi::ModuleInstance;
 use wasmi::RuntimeValue;
 use wasmi::DEFAULT_VALUE_STACK_LIMIT;
 
-use super::state::UpdateCompilationTable;
+
 use super::CompiledImage;
 use super::ExecutionResult;
 
@@ -51,7 +51,6 @@ impl Execution<RuntimeValue>
         let instance = self.instance.run_start(externals).unwrap();
 
         let result = instance.invoke_export(&self.entry, &[], externals)?;
-
         Ok(result)
     }
 
@@ -67,40 +66,46 @@ impl Execution<RuntimeValue>
 
         let result =
             instance.invoke_export_trace(&self.entry, &[], externals, self.tracer.clone())?;
+        
+        let mut tables: Option<Tables> = None;
+        let tracer = self.tracer.borrow();
+        if !tracer.has_dumped() {
+            let execution_tables = {
+                let tracer = self.tracer.borrow();
+    
+                ExecutionTable {
+                    etable: tracer.etable.clone(),
+                    jtable: Arc::new(tracer.jtable.clone()),
+                }
+            };
+    
+            let updated_init_memory_table = self
+                .tables
+                .update_init_memory_table(&execution_tables.etable.entries());
+    
+            let post_image_table = {
+                CompilationTable {
+                    itable: self.tables.itable.clone(),
+                    imtable: updated_init_memory_table,
+                    elem_table: self.tables.elem_table.clone(),
+                    configure_table: self.tables.configure_table.clone(),
+                    static_jtable: self.tables.static_jtable.clone(),
+                    initialization_state: self
+                        .tables
+                        .update_initialization_state(&execution_tables.etable.entries(), true),
+                }
+            };
 
-        let execution_tables = {
-            let tracer = self.tracer.borrow();
-
-            ExecutionTable {
-                etable: tracer.etable.clone(),
-                jtable: Arc::new(tracer.jtable.clone()),
-            }
-        };
-
-        let updated_init_memory_table = self
-            .tables
-            .update_init_memory_table(&execution_tables.etable.entries());
-
-        let post_image_table = {
-            CompilationTable {
-                itable: self.tables.itable.clone(),
-                imtable: updated_init_memory_table,
-                elem_table: self.tables.elem_table.clone(),
-                configure_table: self.tables.configure_table.clone(),
-                static_jtable: self.tables.static_jtable.clone(),
-                initialization_state: self
-                    .tables
-                    .update_initialization_state(&execution_tables.etable.entries(), true),
-            }
-        };
-
-        Ok(ExecutionResult {
-            tables: Tables {
+            tables = Some(Tables {
                 compilation_tables: self.tables,
                 execution_tables,
                 post_image_table,
                 is_last_slice: true,
-            },
+            });
+        } 
+
+        Ok(ExecutionResult {
+            tables,
             result,
             public_inputs_and_outputs: wasm_io.public_inputs_and_outputs.borrow().clone(),
             outputs: wasm_io.outputs.borrow().clone(),
@@ -121,51 +126,15 @@ impl WasmiRuntime {
         host_plugin_lookup: &HashMap<usize, HostFunctionDesc>,
         entry: &str,
         phantom_functions: &Vec<String>,
+        callback: Option<impl FnMut(Tables, usize) +'static>,
+        capability: usize
     ) -> Result<CompiledImage<wasmi::NotStartedModuleRef<'a>, wasmi::tracer::Tracer>> {
-        let tracer = wasmi::tracer::Tracer::new(host_plugin_lookup.clone(), phantom_functions);
+        let tracer = wasmi::tracer::Tracer::new(host_plugin_lookup.clone(), phantom_functions, callback, capability);
         let tracer = Rc::new(RefCell::new(tracer));
-
         let instance = ModuleInstance::new(&module, imports, Some(tracer.clone()))
             .expect("failed to instantiate wasm module");
 
-        let fid_of_entry = {
-            let idx_of_entry = instance.lookup_function_by_name(tracer.clone(), entry);
-
-            tracer
-                .clone()
-                .borrow_mut()
-                .static_jtable_entries
-                .push(StaticFrameEntry {
-                    enable: true,
-                    frame_id: 0,
-                    next_frame_id: 0,
-                    callee_fid: idx_of_entry,
-                    fid: 0,
-                    iid: 0,
-                });
-
-            if instance.has_start() {
-                tracer
-                    .clone()
-                    .borrow_mut()
-                    .static_jtable_entries
-                    .push(StaticFrameEntry {
-                        enable: true,
-                        frame_id: 0,
-                        next_frame_id: 0,
-                        callee_fid: 0, // the fid of start function is always 0
-                        fid: idx_of_entry,
-                        iid: 0,
-                    });
-            }
-
-            if instance.has_start() {
-                0
-            } else {
-                idx_of_entry
-            }
-        };
-
+        let fid_of_entry = tracer.clone().borrow().get_fid_of_entry();
         let itable = Arc::new(tracer.borrow().itable.clone());
         let imtable = tracer.borrow().imtable.finalized();
         let elem_table = Arc::new(tracer.borrow().elem_table.clone());
